requirements.txt
amulet-core==1.9.32
amulet-leveldb==1.0.2
amulet-nbt==2.1.5
anvil-parser==0.9.0
asttokens==3.0.1
build==1.3.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
frozendict==2.4.7
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
lz4==4.4.5
matplotlib-inline==0.2.1
mutf8==1.0.6
NBT==1.5.1
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==3.11.0
portalocker==2.10.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyMCTranslate==1.2.36
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
uv==0.9.11
wcwidth==0.2.14
wheel==0.45.1
